+++ Backend/server/config.py
@@ -1,10 +1,8 @@
 from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
 from flask_bcrypt import Bcrypt
 from flask_migrate import Migrate
 from flask_restful import Api
-from flask_cors import CORS
-from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy import MetaData
 
 
 app = Flask(__name__)
@@ -13,15 +11,10 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///pesabank.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 app.json.compact = False
 
-metadata = MetaData(naming_convention={
-    "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
-})
-
-db = SQLAlchemy(metadata=metadata)
+db = SQLAlchemy()
 migrate = Migrate(app, db)
 db.init_app(app)
 
 bcrypt = Bcrypt(app)
 
 api = Api(app)
-CORS(app)
