--- tested-server(partial)/Pipfile
@@ -1,21 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-flask = "2.2.2"
-flask-sqlalchemy = "3.0.3"
-Werkzeug = "2.2.2"
-sqlalchemy-serializer = "1.4.1"
-flask-migrate = "4.0.0"
-flask-restful = "0.3.9"
-importlib-metadata = "6.0.0"
-importlib-resources = "5.10.0"
-sqlalchemy = "2.0.25"
-flask-bcrypt = "1.0.1"
-
-[dev-packages]
-
-[requires]
-python_full_version = "3.8.13"
